BIG IMPROVEMENT
Now the function update_paging uses the info on the
image side only if it already loaded, this way on the
main thread there aren't any blocking calls waiting on the
thread pool.
Because of this change, update_paging is now called periodically
until all the image sides of a tag are loaded.
Also now the image loading of the current image is done on the
main thread.